+++ super-mario-3-se102/GameSettings.h
@@ -2,11 +2,11 @@
 #include "Utils.h"
 
 struct GameSettings {
-	Utils::Dimension<float> gameDimension;
+	Utils::Dimension<int> gameDimension;
 	int pixelScale;
 	int maxFPS;
 	bool pixelPerfectRendering;
 	std::string dataDirectory;
-	GameSettings(const Utils::Dimension<float>& gameDimension, int pixelScale, int maxFPS, bool pixelPerfectRendering, const std::string& dataDirectory);
+	GameSettings(const Utils::Dimension<int>& gameDimension, int pixelScale, int maxFPS, bool pixelPerfectRendering, const std::string& dataDirectory);
 	GameSettings();
 };
