+++ super-mario-3-se102/Grid.h
@@ -21,21 +21,21 @@ public:
 	typedef EntitiesInCell* LPEntitiesInCell;
 	typedef const EntitiesInCell* LPConstEntitiesInCell;
 
-	Grid(int numberOfColumns, int numberOfRows, const Utils::Dimension<float>& cellSize);
+	Grid(int numberOfColumns, int numberOfRows, const Utils::Dimension<int>& cellSize);
 	~Grid();
 	virtual void AddToCell(LPEntity entity, const Utils::Vector2<int>& cellIndex);
 	LPConstEntitiesInCell EntitiesAt(const Utils::Vector2<int>& cellIndex);
 	void ForEachEntityIn(const CellRange& range, std::function<void(LPEntity)> handler);
 	void ForEachEntity(std::function<void(LPEntity)> handler);
 	Utils::Vector2<int> GetCellIndexAtPoint(const Utils::Vector2<float>& point);
-	CellRange GetCellRangeFromRectangle(const Utils::Vector2<float>& position, const Utils::Dimension<float>& dim);
+	CellRange GetCellRangeFromRectangle(const Utils::Vector2<float>& position, const Utils::Dimension<int>& dim);
 	void OnEntityDestroy(LPEntity entity);
 protected:
 	//each cell contains a list of entities
 	std::vector<LPEntitiesInCell> cells;
 	int numOfRows;
 	int numOfCols;
-	Utils::Dimension<float> cellSize;
+	Utils::Dimension<int> cellSize;
 };
 typedef Grid* LPGrid;
