+++ super-mario-3-se102/SceneLoader.h
@@ -12,7 +12,7 @@ class SceneLoader
 public:
 	static LPScene LoadScene(std::string scenePath);
 private:
-	static std::string ParseWorldProperties(std::ifstream& file, std::string& sceneType, Utils::Dimension<float>& dim, D3DCOLOR& bgColor, std::string& prevScenePath);
+	static std::string ParseWorldProperties(std::ifstream& file, std::string& sceneType, Utils::Dimension<int>& dim, D3DCOLOR& bgColor, std::string& prevScenePath);
 	static std::string ParseEncodedWorld(std::ifstream& file, int world_width, const std::string& sceneType, LPEncodedWorld& encodedWorld);
 	static std::string ParseSpatialPartitionGrid
 	(std::ifstream& file, LPGrid& wallEntitySpatialGrid, LPGrid& staticEntitySpatialGrid, LPDynamicGrid& movableEntitySpatialGrid);
